feat: fix and improve rate limit handling. (#176)
* feat: fix and improve retries and rate limit handling.

The SDK now retries on network errors and the default retry handling has been fixed
for both the calls to the OpenFGA API and the API Token Issuer for those using ClientCredentials

The SDK now also respects the rate limit headers (`Retry-After`) returned by the server and will retry the request after the specified time.
If the header is not sent or on network errors, it will fall back to exponential backoff.

* chore: fix changelog